+++ src/main/java/finalmission/member/service/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package finalmission.member.service.dto;
+package finalmission.member.service.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
