+++ src/main/java/finalmission/member/service/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package finalmission.member.service.dto;
+package finalmission.member.service.dto.response;
 
 public record LoginResponse(
         String token
