+++ src/main/java/finalmission/reservation/controller/ReservationController.java
@@ -3,11 +3,11 @@ package finalmission.reservation.controller;
 import finalmission.auth.MemberSubject;
 import finalmission.auth.RequiredOwner;
 import finalmission.reservation.service.ReservationService;
-import finalmission.reservation.service.dto.CreateReservationInformationRequest;
-import finalmission.reservation.service.dto.ReservationDetailResponse;
-import finalmission.reservation.service.dto.ReservationInformationResponse;
-import finalmission.reservation.service.dto.CreateReservationRequest;
-import finalmission.reservation.service.dto.ReservationResponse;
+import finalmission.reservation.service.dto.request.CreateReservationInformationRequest;
+import finalmission.reservation.service.dto.response.ReservationDetailResponse;
+import finalmission.reservation.service.dto.response.ReservationInformationResponse;
+import finalmission.reservation.service.dto.request.CreateReservationRequest;
+import finalmission.reservation.service.dto.response.ReservationResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
