+++ src/main/java/finalmission/reservation/service/ReservationService.java
@@ -6,11 +6,11 @@ import finalmission.reservation.domain.Reservation;
 import finalmission.reservation.domain.ReservationInformation;
 import finalmission.reservation.repository.ReservationInformationRepository;
 import finalmission.reservation.repository.ReservationRepository;
-import finalmission.reservation.service.dto.CreateReservationInformationRequest;
-import finalmission.reservation.service.dto.ReservationDetailResponse;
-import finalmission.reservation.service.dto.ReservationInformationResponse;
-import finalmission.reservation.service.dto.CreateReservationRequest;
-import finalmission.reservation.service.dto.ReservationResponse;
+import finalmission.reservation.service.dto.request.CreateReservationInformationRequest;
+import finalmission.reservation.service.dto.response.ReservationDetailResponse;
+import finalmission.reservation.service.dto.response.ReservationInformationResponse;
+import finalmission.reservation.service.dto.request.CreateReservationRequest;
+import finalmission.reservation.service.dto.response.ReservationResponse;
 import finalmission.restaurant.domain.Restaurant;
 import finalmission.restaurant.repository.RestaurantRepository;
 import org.springframework.stereotype.Service;
