+++ src/test/java/finalmission/reservation/service/ReservationServiceTest.java
@@ -4,9 +4,9 @@ import finalmission.member.domain.Member;
 import finalmission.member.domain.vo.Role;
 import finalmission.reservation.domain.Reservation;
 import finalmission.reservation.domain.ReservationInformation;
-import finalmission.reservation.service.dto.CreateReservationInformationRequest;
-import finalmission.reservation.service.dto.CreateReservationRequest;
-import finalmission.reservation.service.dto.ReservationDetailResponse;
+import finalmission.reservation.service.dto.request.CreateReservationInformationRequest;
+import finalmission.reservation.service.dto.request.CreateReservationRequest;
+import finalmission.reservation.service.dto.response.ReservationDetailResponse;
 import finalmission.restaurant.domain.Restaurant;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
